extracted getter-buff-reducer into separate file
